Added executable to run with profilers
- Nvvp and Nsight seem to not work with IO to stdin/stdout, so we add a version that takes input and output file names as flags.
- Added options to profile bicgstab and gmres, though not all options are exposed yet.
